modelo-arima-streamlit/requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.23.5
pmdarima==2.0.3
statsmodels==0.14.0
plotly==5.15.0
google-cloud-storage==2.10.0
google-auth==2.22.0
joblib